Add reusable pagination component
