+++ src/views/MobilitySettingsView/components/RouteList/RouteList.js
@@ -3,14 +3,12 @@ import { Checkbox, FormControlLabel, Typography } from '@mui/material';
 import PropTypes from 'prop-types';
 import useLocaleText from '../../../../utils/useLocaleText';
 import { isDataValid } from '../../../../components/MobilityPlatform/utils/utils';
-import SMButton from '../../../../components/ServiceMapButton';
-import Container from '../../../../components/Container';
 import RouteLength from '../RouteLength';
 import Description from '../Description';
+import Pagination from '../Pagination';
 
 const RouteList = ({
   classes,
-  intl,
   openList,
   items,
   itemsPerPage,
@@ -69,41 +67,23 @@ const RouteList = ({
       : null;
   };
 
-  const renderPagination = () => {
-    const totalPages = Math.ceil(items.length / itemsPerPage);
-    const paginationLinks = [];
-
-    for (let i = 1; i <= totalPages; i += 1) {
-      paginationLinks.push(
-        <SMButton
-          key={i}
-          aria-label={intl.formatMessage({ id: 'general.pagination.page.number' }, { number: i })}
-          onClick={() => setCurrentPage(i)}
-          className={`${classes.button} ${currentPage === i ? classes.active : ''}`}
-          variant="contained"
-          role="link"
-        >
-          {i}
-        </SMButton>,
-      );
-    }
-
-    return <Container className={classes.pagination}>{paginationLinks}</Container>;
-  };
-
   return (
     <div>
       <div className={classes.listContainer}>{renderList()}</div>
-      {openList ? renderPagination() : null}
+      {openList ? (
+        <Pagination
+          items={items}
+          itemsPerPage={itemsPerPage}
+          currentPage={currentPage}
+          setCurrentPage={setCurrentPage}
+        />
+      ) : null}
     </div>
   );
 };
 
 RouteList.propTypes = {
   classes: PropTypes.objectOf(PropTypes.any).isRequired,
-  intl: PropTypes.shape({
-    formatMessage: PropTypes.func,
-  }).isRequired,
   openList: PropTypes.bool,
   items: PropTypes.arrayOf(PropTypes.any),
   itemsPerPage: PropTypes.number,
