+++ src/views/MobilitySettingsView/components/RouteList/index.js
@@ -1,6 +1,5 @@
 import { withStyles } from '@mui/styles';
-import { injectIntl } from 'react-intl';
 import RouteList from './RouteList';
 import styles from './styles';
 
-export default withStyles(styles)(injectIntl(RouteList));
+export default withStyles(styles)(RouteList);
