+++ src/views/MobilitySettingsView/components/TrailList/index.js
@@ -1,6 +1,5 @@
 import { withStyles } from '@mui/styles';
-import { injectIntl } from 'react-intl';
 import TrailList from './TrailList';
 import styles from './styles';
 
-export default withStyles(styles)(injectIntl(TrailList));
+export default withStyles(styles)(TrailList);
